Created MAL scraper
Created the file mal_scrape.py which contains the scraper for the MAL website.
It downloads the anime images from the site and ignores all meta site images.

Also adjusted the .gitignore to not include downloaded images and the data folder.
This is to reduce the size of the repository and allow for easier uploading to github.