Modified rejected image method
Added a bad_file_id.txt file which houses all the names of files I handpicked out of the dataset.
This is not a comprehensive list of all the files in the dataset, but it is a good starting point.
Uses numpy to read in the file and store it in a numpy array, which is converted to a list.